+++ e2e/scripts/get_node_ip.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-docker inspect -f '{{ .NetworkSettings.Networks.monet.IPAddress }}' $1
-
-