+++ e2e/scripts/node_restart.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-for i in "$@"
-do
-  docker restart $i 
-done