+++ src/service/service.go
@@ -1,14 +1,10 @@
 package service
 
 import (
-	"io/ioutil"
 	"net/http"
 	"os"
-	"strings"
 	"sync"
 
-	"github.com/ethereum/go-ethereum/accounts/keystore"
-	"github.com/gorilla/mux"
 	"github.com/mosaicnetworks/evm-lite/src/state"
 	"github.com/sirupsen/logrus"
 )
@@ -19,35 +15,27 @@ type infoCallback func() (map[string]string, error)
 
 type Service struct {
 	sync.Mutex
-	state       *state.State
-	submitCh    chan []byte
-	keystoreDir string
-	apiAddr     string
-	keyStore    *keystore.KeyStore
-	pwdFile     string
-	getInfo     infoCallback
-	logger      *logrus.Logger
+	state    *state.State
+	submitCh chan []byte
+	apiAddr  string
+	getInfo  infoCallback
+	logger   *logrus.Entry
 }
 
-func NewService(keystoreDir, apiAddr, pwdFile string,
+func NewService(apiAddr string,
 	state *state.State,
 	submitCh chan []byte,
-	logger *logrus.Logger) *Service {
+	logger *logrus.Entry) *Service {
+
 	return &Service{
-		keystoreDir: keystoreDir,
-		apiAddr:     apiAddr,
-		pwdFile:     pwdFile,
-		state:       state,
-		submitCh:    submitCh,
-		logger:      logger}
+		apiAddr:  apiAddr,
+		state:    state,
+		submitCh: submitCh,
+		logger:   logger}
 }
 
 func (m *Service) Run() {
-	m.checkErr(m.makeKeyStore())
-
-	m.checkErr(m.unlockAccounts())
-
-	m.logger.Info("serving api...")
+	m.logger.WithField("bind_address", m.apiAddr).Info("API")
 	m.serveAPI()
 }
 
@@ -59,81 +47,26 @@ func (m *Service) SetInfoCallback(f infoCallback) {
 	m.getInfo = f
 }
 
-func (m *Service) makeKeyStore() error {
-
-	scryptN := keystore.StandardScryptN
-	scryptP := keystore.StandardScryptP
-
-	if err := os.MkdirAll(m.keystoreDir, 0700); err != nil {
-		return err
-	}
-
-	m.keyStore = keystore.NewKeyStore(m.keystoreDir, scryptN, scryptP)
-
-	return nil
-}
-
-func (m *Service) unlockAccounts() error {
-
-	if len(m.keyStore.Accounts()) == 0 {
-		return nil
-	}
-
-	pwd, err := m.readPwd()
-	if err != nil {
-		m.logger.WithError(err).Error("Reading PwdFile")
-		return err
-	}
-
-	for _, ac := range m.keyStore.Accounts() {
-		if err := m.keyStore.Unlock(ac, string(pwd)); err != nil {
-			return err
-		}
-		m.logger.WithField("address", ac.Address.Hex()).Debug("Unlocked account")
-	}
-	return nil
-}
-
+// Serve registers the API handlers with the DefaultServerMux of the http
+// package, and calls ListenAndServe. It is possible that another module in the
+// application (ex: the consensus system) has registered other handlers with the
+// DefaultServeMux. In this case, those handlers will also be process by this
+// server.
 func (m *Service) serveAPI() {
+	// Add handlers to DefaultServerMux
+	http.HandleFunc("/account/", m.makeHandler(accountHandler))
+	http.HandleFunc("/call", m.makeHandler(callHandler))
+	http.HandleFunc("/rawtx", m.makeHandler(rawTransactionHandler))
+	http.HandleFunc("/tx/", m.makeHandler(transactionReceiptHandler))
+	http.HandleFunc("/info", m.makeHandler(infoHandler))
+	http.HandleFunc("/poa", m.makeHandler(poaHandler))
+	http.HandleFunc("/genesis", m.makeHandler(genesisHandler))
 
-	serverMuxEVM := http.NewServeMux()
-
-	r := mux.NewRouter()
-	r.HandleFunc("/account/{address}", m.makeHandler(accountHandler)).Methods("GET")
-	r.HandleFunc("/accounts", m.makeHandler(accountsHandler)).Methods("GET")
-	r.HandleFunc("/call", m.makeHandler(callHandler)).Methods("POST")
-	r.HandleFunc("/tx", m.makeHandler(transactionHandler)).Methods("POST")
-	r.HandleFunc("/rawtx", m.makeHandler(rawTransactionHandler)).Methods("POST")
-	r.HandleFunc("/tx/{tx_hash}", m.makeHandler(transactionReceiptHandler)).Methods("GET")
-	r.HandleFunc("/info", m.makeHandler(infoHandler)).Methods("GET")
-	r.HandleFunc("/html/info", m.makeHandler(htmlInfoHandler)).Methods("GET")
-	r.HandleFunc("/contract", m.makeHandler(contractHandler)).Methods("GET")
-	r.HandleFunc("/poa", m.makeHandler(poaHandler)).Methods("GET")
-	r.HandleFunc("/genesis", m.makeHandler(genesisHandler)).Methods("GET")
-
-	serverMuxEVM.Handle("/", &CORSServer{r})
-
-	m.logger.WithField("apiAddr", m.apiAddr).Debug("EVM-Lite Service serving")
-	http.ListenAndServe(m.apiAddr, serverMuxEVM)
-}
-
-type CORSServer struct {
-	r *mux.Router
-}
-
-func (s *CORSServer) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
-	if origin := req.Header.Get("Origin"); origin != "" {
-		rw.Header().Set("Access-Control-Allow-Origin", origin)
-		rw.Header().Set("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
-		rw.Header().Set("Access-Control-Allow-Headers",
-			"Accept, Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization")
+	// The call to ListenAndServe is a blocking operation
+	err := http.ListenAndServe(m.apiAddr, nil)
+	if err != nil {
+		m.logger.Error(err)
 	}
-	// Stop here if its Preflighted OPTIONS request
-	if req.Method == "OPTIONS" {
-		return
-	}
-	// Lets Gorilla work
-	s.r.ServeHTTP(rw, req)
 }
 
 func (m *Service) makeHandler(fn func(http.ResponseWriter, *http.Request, *Service)) http.HandlerFunc {
@@ -149,17 +82,4 @@ func (m *Service) checkErr(err error) {
 		m.logger.WithError(err).Error("ERROR")
 		os.Exit(1)
 	}
-}
-
-func (m *Service) readPwd() (pwd string, err error) {
-	text, err := ioutil.ReadFile(m.pwdFile)
-	if err != nil {
-		return "", err
-	}
-	lines := strings.Split(string(text), "\n")
-	// Sanitise DOS line endings.
-	for i := range lines {
-		lines[i] = strings.TrimRight(lines[i], "\r")
-	}
-	return lines[0], nil
 }