+++ src/state/was.go
@@ -34,7 +34,7 @@ type WriteAheadState struct {
 	totalUsedGas uint64
 	gp           *core.GasPool
 
-	logger *logrus.Logger
+	logger *logrus.Entry
 }
 
 func NewWriteAheadState(db ethdb.Database,
@@ -43,7 +43,7 @@ func NewWriteAheadState(db ethdb.Database,
 	chainConfig params.ChainConfig,
 	vmConfig vm.Config,
 	gasLimit uint64,
-	logger *logrus.Logger) (*WriteAheadState, error) {
+	logger *logrus.Entry) (*WriteAheadState, error) {
 
 	ethState, err := ethState.New(root, ethState.NewDatabase(db))
 	if err != nil {
@@ -131,6 +131,12 @@ func (was *WriteAheadState) ApplyTransaction(tx ethTypes.Transaction, txIndex in
 }
 
 func (was *WriteAheadState) Commit() (common.Hash, error) {
+	was.logger.WithFields(logrus.Fields{
+		"txs":      was.txIndex,
+		"receipts": len(was.receipts),
+		"logs":     len(was.allLogs),
+	}).Info("Commit")
+
 	// Commit all state changes to the database
 	root, err := was.ethState.Commit(true)
 	if err != nil {